Remove duplicates from RefSeq GRCh37 cache
The RefSeq cache is built from the otherfeatures database.
This contains two imports from GFF, one curated and
one not. On the frozen GRCh37 DB the external references
are attached to the "wrong" set, meaning VEP has to copy
references between the two and then retain only the
transcript from the "correct" set. Since the database
will not be updated this hack will remain in place.