Use cached seq_region_start and seq_region_end
Historically VEP has used $vf->{start} for coordinate lookup, and
mostly it works fine. Some of the TranscriptMapper methods
require $vf->seq_region_start(). Mostly this is OK as it just does
a slice-based lookup, but for circular slices e.g. in E.coli the
core API tries to do a DB lookup based on $vf->dbID, which
of course fails.

We get around it by caching seq_region_start on the VF itself.
This depends on an update the ensembl-variation
BaseVariationFeature module to implement that caching and avoid
defaulting to the parent method in Feature.

Have to be careful though, particularly when split_variants() and
minimise_alleles() create copies of VFs, hence the changes there.