Redefine specified $config->{cache} to be "1"
This ensures that the behavior of `--cache` and `--cache BLAH` are the same downstream.  

The original behavior in master was that `--cache` set `$config->{cache} == 1`, and if omitted, it was undefined.  

This causes that behavior to return exactly as before, but throws a warning if the optional parameter is not a directory.